+++ draft-packages/tooltip/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.7.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tooltip@3.7.7...@kaizen/draft-tooltip@3.7.8) (2022-03-07)
+
+**Note:** Version bump only for package @kaizen/draft-tooltip
+
+
+
+
+
+## [3.7.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tooltip@3.7.6...@kaizen/draft-tooltip@3.7.7) (2022-03-07)
+
+**Note:** Version bump only for package @kaizen/draft-tooltip
+
+
+
+
+
+## [3.7.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tooltip@3.7.5...@kaizen/draft-tooltip@3.7.6) (2022-03-07)
+
+**Note:** Version bump only for package @kaizen/draft-tooltip
+
+
+
+
+
 ## [3.7.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tooltip@3.7.4...@kaizen/draft-tooltip@3.7.5) (2022-03-01)
 
 **Note:** Version bump only for package @kaizen/draft-tooltip